+++ modules/aws/eks/eks.tf
@@ -13,6 +13,8 @@ locals {
     token          = "foobar"
     calico_version = local.calico_version
   })
+  node_labels = { for np in var.eks_config.node_pools : np.name => { for key, value in np.node_labels : "k8s.io/cluster-autoscaler/node-template/label/${key}" => value } }
+  node_taints = { for np in var.eks_config.node_pools : np.name => { for node_taint in np.node_taints : "k8s.io/cluster-autoscaler/node-template/taint/${node_taint["key"]}" => "${node_taint["value"]}:${node_taint["effect"]}" } }
 }
 
 data "aws_subnet" "cluster" {
@@ -208,7 +210,11 @@ resource "aws_eks_node_group" "this" {
 
   labels = merge({ "xkf.xenit.io/node-ttl" = "168h" }, each.value.node_labels, { "xkf.xenit.io/node-pool" = each.value.name })
 
-  tags = local.global_tags
+  tags = merge(
+    local.global_tags,
+    local.node_labels[each.key],
+    local.node_taints[each.key],
+  )
 
   dynamic "taint" {
     for_each = each.value.node_taints